Type fixes!

+++ Definitions/Prices.py
@@ -4,7 +4,7 @@
 class DefaultPrices:
     """Default Prices for things in Shop"""
 
-    Rent: int = 100.00
+    Rent: float = 100.00
     NewWorker: float = 500.00
     Salary: float = 8.00
     Table: float = 250.00
